Move dataset management Python files to their own folder.
